requirements.txt
pandas
scikit-learn
matplotlib
seaborn
joblib
streamlit
numpy
Pillow